requirements.txt
django==2.2
django-taggit==1.3.0
pillow==8.2.0

